pyproject.toml
[tool.poetry]
name = "test_project"
version = "0.1.0"
description = ""
authors = ["Egor Loskutov <egorkaloskutov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
channels = "^4.0.0"
django-filter = "^23.5"
djoser = "^2.2.2"
pillow = "^10.2.0"
daphne = "^3.0.0"
channels-redis = "^4.2.0"
selenium = "^4.16.0"
pyjwt = "^2.8.0"
djangochannelsrestframework = "^1.2.0"
gunicorn = "^21.2.0"
redis-server = "^6.0.9"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
asgiref = "^3.7.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"